+++ aldrin/src/lib.rs
@@ -86,6 +86,7 @@ pub use aldrin_macros::IntrospectableFromAldrin as Introspectable;
 #[cfg(feature = "codegen")]
 pub use aldrin_macros::{
     generate, DeserializeFromAldrin as Deserialize, SerializeFromAldrin as Serialize,
+    SerializeKeyFromAldrin as SerializeKey,
 };
 pub use bus_listener::BusListener;
 pub use channel::{