+++ core/src/lib.rs
@@ -31,7 +31,7 @@ pub mod transport;
 #[cfg(all(feature = "derive", feature = "introspection"))]
 pub use aldrin_macros::Introspectable;
 #[cfg(feature = "derive")]
-pub use aldrin_macros::{Deserialize, Serialize};
+pub use aldrin_macros::{Deserialize, Serialize, SerializeKey};
 pub use bus_listener::{BusEvent, BusListenerFilter, BusListenerScope, BusListenerServiceFilter};
 pub use channel_end::{ChannelEnd, ChannelEndWithCapacity};
 pub use deserialize_key::{DeserializeKey, DeserializeKeyImpl};