+++ macros/src/derive.rs
@@ -2,9 +2,11 @@ mod deserialize;
 mod introspectable;
 mod options;
 mod serialize;
+mod serialize_key;
 
 use options::{ItemOptions, Options};
 
 pub use deserialize::{gen_deserialize_from_aldrin, gen_deserialize_from_core};
 pub use introspectable::{gen_introspectable_from_aldrin, gen_introspectable_from_core};
 pub use serialize::{gen_serialize_from_aldrin, gen_serialize_from_core};
+pub use serialize_key::{gen_serialize_key_from_aldrin, gen_serialize_key_from_core};